Use edit_resource on yum repos if they exist
This fixes problems when using this in production where the yum, yum-centos,
yum-epel and yum-centos recipes might be included in a community cookbook. If
this happens, then the resources in this cookbook will fail to work and go back
to the default settings.

By doing it this way, it ensures we get it correctly. This uses a similar logic
used in the upstream cookbooks, but using edit_resource instead.