Fix issues with yum_repository resources not working
This works around an issue we're running into where the various yum_repository
resources we edit are not properly being run in the order we expect. Instead of
using edit_resource, we should be using declare_resource which will ensure the
resource is executed at the time we call these custom resources.