Use declare_resource correctly
This properly uses declare_resource by doing the following:

- Move all parameter settings we need changed to node.run_state [1] so we can
  loop _only_ the parameters we need changed easily
- Find the specific yum_repository resource and change each parameter we set in
  node.run_state. This ensures we pull any other edit_resource changes to these
  resources at compile time and pull them in.
- Declare the whole resource again looping through all known parameters for
  yum_repository
- Set all node.default attributes we had set in node.run_state so it can be used
  with the recipes if the resource hasn't been defined.

This basically resolves an issue where changes made to the resource after this
gets called gets changed again and causes idempotency issues.

Of course, most of this can go away if proper resources for each repositories
are created but this is unfortunately what we need to do to work around
attributes.

[1] https://docs.chef.io/recipes/#noderun_state